Auto_Test/requirements.txt
pip~=23.2.1
pytest~=7.4.0
selenium~=4.4.0
PyYAML~=6.0.1
allure-pytest~=2.13.2
allure-python-commons~=2.13.2
webdriver-manager~=4.0.0
PyVirtualDisplay~=3.0
Pillow~=10.2.0
ImageHash~=4.3.1
xvfbwrapper
opencv-python